refactor: Extract findTrackedTasks and findUntrackedTasks to index-utils module
